refactored events.jsx
