+++ pages/events/index.js
@@ -1,0 +1,2 @@
+//TOOD clean up indirection
+module.exports = require('./EventsPage.jsx');
